isochron: fix build with mips64 and ppc64
In a strange turn of events, the __s64 and __u64 types exported by the
Linux kernel headers cannot be portably printed directly, neither using
%lld and %llu (as we currently attempt), nor using PRId64 and PRIu64
(which are supposed to solve this, but for the int64_t and uint64_t types).

We have:

arch       __s64 evaluates to    PRId64 evaluates to
powerpc64  long int              lld
x86_64     long long int         ld

To reconcile both kinds of architectures, we need to explicitly cast
to the larger type and keep printing with the larger printf format
specifier (PRId64/PRIu64 aren't useful options).

This fixes multiple issues such as below (for powerpc64), not including
them all for brevity:

src/orchestrate.c: In function ‘isochron_node_rtt_finalize’:
src/orchestrate.c:85:73: error: format ‘%lld’ expects argument of type ‘long long int’, but argument 4 has type ‘__s64’ {aka ‘long int’} [-Werror=format=]
  printf("Max TCP round trip time to node %s over %zu measurements is %lld ns\n",
                                                                      ~~~^
                                                                      %ld
         node->name, node->num_rtt_measurements, node->max_rtt);
                                                 ~~~~~~~~~~~~~
cc1: all warnings being treated as errors

Thanks to Bastian Germann <bage@debian.org> for reporting this to me.

+++ src/orchestrate.c
@@ -83,7 +83,7 @@ static void isochron_node_rtt_after(struct isochron_orch_node *node)
 static void isochron_node_rtt_finalize(struct isochron_orch_node *node)
 {
 	printf("Max TCP round trip time to node %s over %zu measurements is %lld ns\n",
-	       node->name, node->num_rtt_measurements, node->max_rtt);
+	       node->name, node->num_rtt_measurements, (long long)node->max_rtt);
 }
 
 static int prog_query_test_state(struct isochron_orch_node *node,
